# PeAPIniere/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PéAPInière API",
        "description": "Documentation de l'API pour la gestion des Pépinière.",
        "contact": {
            "email": "support@PéAPInière.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/plants": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Obtenir toutes les plantes",
                "description": "Récupérer la liste des plantes.",
                "operationId": "6e2a3a91967b1e364c87b755cd5152f6",
                "responses": {
                    "200": {
                        "description": "Liste des plantes récupérée avec succès"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "post": {
                "tags": [
                    "Plants"
                ],
                "summary": "Créer une plante",
                "description": "Ajouter une nouvelle plante.",
                "operationId": "d005e00370ab9bd321d50ba5dde92042",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Rose"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "rose"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Belle plante"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 19.99
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plante ajoutée avec succès"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/plants/{slug}": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Obtenir une plante par son slug",
                "description": "Obtenir une plante par slug.",
                "operationId": "cf101271037916b9c4a3e8bed0a3bf4e",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug de la plante",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plante trouvée"
                    },
                    "404": {
                        "description": "Plante introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Plants"
                ],
                "summary": "Mettre à jour une plante",
                "description": "Mettre à jour une plante.",
                "operationId": "e5981af9ce55cca7fc8eee04c3c0b488",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la plante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tulipe"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "tulipe"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Plante colorée"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 14.99
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plante mise à jour avec succès"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Plants"
                ],
                "summary": "Supprimer une plante",
                "description": "Supprimer une plante.",
                "operationId": "72abaa41b86e9d3386648d6c79c2b1ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la plante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plante supprimée avec succès"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}